Refactored scenes a bit
