+++ src/feature/scene/index.ts
@@ -1,8 +1,9 @@
-export { default as World, Actor } from '../../foundation/engine/rendering/world'
+export { default as World } from '../../foundation/engine/scene/world'
 export { default as Mesh } from '../../foundation/engine/rendering/mesh'
 export { default as Positionable } from '../../foundation/engine/rendering/positionable'
 export { default as Rotatable } from '../../foundation/engine/rendering/rotatable'
 export { default as Updatable } from '../../foundation/engine/update/updatable'
-export { default as Camera } from '../../foundation/engine/rendering/camera'
+export { default as Camera } from '../../foundation/engine/scene/camera'
 export { default as Coordinate, Coordinate2d, Coordinate3d } from '../../foundation/engine/space/coordinates'
 export { default as Scene } from './scene'
+export { default as Actor } from '../../foundation/engine/scene/actor'
