+++ src/foundation/engine/rendering/render-loop.ts
@@ -1,8 +1,8 @@
 import { Canvas, Context } from './canvas'
 import Scene from '../../../feature/scene/scene'
-import { Actor } from './world'
+import Actor from '../scene/actor'
 import { ShaderInfo, createPositionBuffer, initializeFrameSettings, setDefaultShaders, setPositionAttribute } from './gl'
-import { getProjectionMatrices } from './camera'
+import { getProjectionMatrices } from '../scene/camera'
 import { mat4 } from 'gl-matrix'
 import { isMesh } from './mesh'
 import { isRotatable } from './rotatable'
