--- src/foundation/engine/rendering/world.ts
@@ -1,9 +1,0 @@
-import Unique from '../../base-types/unique'
-
-type ActorContainer = {
-    actors?: () => Actor[]
-}
-
-export type Actor = Unique & ActorContainer
-
-export default interface World extends ActorContainer {}
