+++ src/foundation/engine/scene/overlay.ts
@@ -1,0 +1,8 @@
+import { ActorOptionals } from './actor';
+
+interface OverlayBase {
+
+}
+
+type Overlay = OverlayBase & ActorOptionals
+export default Overlay
