+++ src/foundation/engine/scene/world.ts
@@ -1,0 +1,4 @@
+import { ActorContainer } from './actor'
+
+type World = ActorContainer
+export default World
